+++ vitest.shims.d.ts
@@ -1,5 +1,0 @@
-import type { TestingLibraryMatchers } from '@testing-library/jest-dom/matchers';
-
-declare module 'vitest' {
-  interface Assertion<T = unknown> extends jest.Matchers<void, T>, TestingLibraryMatchers<T, void> {}
-}